// frontend/src/components/Horoscopes/ProcessorProject/src/models/enums/routes.ts
export enum routes {
  astrologicalProcessor = '/processor',
  chats = '/chats',
  authorization = '/authorization',
  settings = '/settings',
  scholar = '/scholar',
  menu = '/menu',
  personal = '/personal',
  myHoroscopes = '/my-horoscopes',
  horoscopes = '/horoscopes',
  forum = '/topics',
  rates = '/rates',
  topicCreator = '/topics/create',
  TopicEditorId = '/topics/:id/edit',
  topic = '/topics/',
  topicId = '/topics/:id',
  notifications = '/notifications',
  calendar = '/calendar',
  main = '/',
  Chat = '/chat',
  userId = '/user/:id',
  user = '/user/',
  userEdit = '/user-edit',
  Courses = '/courses',
  levels = '/levels',
  messages = '/messages',
}