// frontend/src/components/Horoscopes/ProcessorProject/src/models/enums/Notification.ts
export enum NotificationVerb {
  inviteToTopic = 'invite to topic',
  joinToTopic = 'requested to join topic',
  subscribedToUser = 'subscribed to user',
  acceptToTopic = 'accepted to topic',
  commentInTopic = 'comment in topic',
  replyToСomment = 'reply to comment',
  newTopic = 'new topic'
};

export enum NotificationLevel {
  info = 'info',
  success = 'success',
  warning = 'warning',
  error = 'error'
}